Separate out actionAngleTorus tests so they can be easily ignored on appveyor
